ORC/DWRF has separate header for proto metadata
