+++ velox/dwio/dwrf/common/Common.h
@@ -317,72 +317,4 @@ class PostScript {
     WriterVersion writerVersion_;
 };
 
-class DWRFPostScript : public PostScript {
-  public:
-    DWRFPostScript() : PostScript() {}
-    DWRFPostScript(
-      uint64_t footerLength,
-      proto::CompressionKind compression,
-      uint64_t compressionBlockSize,
-      uint32_t writerVersion,
-      proto::StripeCacheMode cacheMode,
-      uint32_t cacheSize)
-    : PostScript(footerLength, static_cast<CompressionKind>(compression),
-        compressionBlockSize, writerVersion)
-    , cacheMode_{static_cast<StripeCacheMode>(cacheMode)}
-    , cacheSize_{cacheSize} {
-    }
-
-    StripeCacheMode cachemode() const {
-      return cacheMode_;
-    }
-
-    bool has_cachemode() const {
-      return cacheMode_ != StripeCacheMode::NA;
-    }
-
-    uint32_t cachesize() const {
-      return cacheSize_;
-    }
-    
-    bool has_cachesize() const {
-      return cacheSize_ != 0;
-    }
-
-  private:
-    StripeCacheMode cacheMode_;
-    uint32_t cacheSize_;
-};
-
-// TODO: move elsewhere
-static CompressionKind convertCompressionKind(proto::orc::CompressionKind
-    compression) {
-  auto compressionUint = static_cast<uint32_t>(compression);
-  if (compressionUint >= 4 && compressionUint <= 5) {
-    compressionUint = 9 - compressionUint;
-  }
-  return static_cast<CompressionKind>(compressionUint);
-}
-
-class ORCPostScript : public PostScript {
-  public:
-    ORCPostScript(
-      uint64_t footerLength,
-      proto::orc::CompressionKind compression,
-      uint64_t compressionBlockSize,
-      uint64_t metadataLength,
-      uint32_t writerVersion,
-      uint64_t stripeStatisticsLength)
-    : PostScript(footerLength, convertCompressionKind(compression),
-        compressionBlockSize, writerVersion)
-    , metadataLength_{metadataLength}
-    , stripeStatisticsLength_{stripeStatisticsLength} {
-    }
-
-  private:
-    uint64_t metadataLength_;
-    uint64_t stripeStatisticsLength_;
-};
-
-
 } // namespace facebook::velox::dwrf
